pyproject.toml
[project]
name = "rag-knowledge-engine"
version = "1.0.0"
description = "Educational RAG pipeline with LangChain, ChromaDB, and swappable LLM providers (OpenAI + Anthropic)"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Mathieu Calvo"}]
readme = "README.md"

dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-chroma>=0.1.0",
    "langchain-community>=0.3.0",
    "chromadb>=0.5.0",
    "openai>=1.40.0",
    "anthropic>=0.34.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "sentence-transformers>=3.0.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    "scikit-learn>=1.4.0",
]

[project.optional-dependencies]
app = ["streamlit>=1.38.0"]
eval = ["ragas>=0.2.0", "datasets>=2.0.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "nbstripout>=0.7.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "requires_api_key: tests that require API keys (deselect with '-m not requires_api_key')",
]
